Enforce uniqueness in steam of events
When we have it happening on an interval, we may get a query that pulls
before downstream processing has finished. We don't want to re-process
something because then we'll be inconsistent, so filter out any
duplicates by observing the stream.

This also enables passing in a max id to query the batch  from, which
avoids the need to reference the materializer state while querying
events, which enables more separation between the materialized states
and the event source.